backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
schwab-py>=1.4.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
